Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,8 +2,8 @@
 #include<stdio.h>
 int main(){
     int a,b,c;
-    int avg;
+    float avg;
     scanf("%d %d %d",&a,&b,&c);
     avg=a+b+c/3;
-    printf("Average: %.2d",avg);
+    printf("Average: %.2f",avg);
 }
